+++ CloudHarvestApi/blueprints/base.py
@@ -1,8 +1,11 @@
 from CloudHarvestCoreTasks.cache import CachedData
+from CloudHarvestCoreTasks.silos import BaseSilo
 
 from flask import Request, Response, jsonify
+from logging import getLogger
 from typing import Any
 
+logger = getLogger('harvest')
 
 ########################################################################################################################
 # FUNCTIONS
@@ -36,6 +39,62 @@ def safe_jsonify(success: bool, reason: str, result: Any, default: Any = None) -
         })
 
     return try_result
+
+class RedisRequest:
+    def __init__(self, silo: str or BaseSilo, max_attempts: int = 10):
+
+        self.silo = silo
+        self.max_attempts = max_attempts
+
+        self.client = None
+
+    def __enter__(self):
+        """
+        Context manager to enter the RedisRequest.
+        """
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        """
+        Context manager to exit the RedisRequest.
+        """
+        # No specific cleanup needed for RedisRequest
+        if self.client:
+            try:
+                self.client.close()
+            except Exception as e:
+                pass
+        pass
+
+    def __getattr__(self, name):
+        """
+        Dynamically wrap StrictRedis methods with retry logic.
+        """
+
+        def wrapper(*args, **kwargs):
+            from CloudHarvestCoreTasks.silos import get_silo
+            self.silo = get_silo(self.silo) if isinstance(self.silo, str) else self.silo
+
+            for i in range(self.max_attempts):
+                try:
+                    self.client = self.silo.connect()
+
+                    result = getattr(self.client, name)(*args, **kwargs)
+                    return result
+
+                except BaseException as ex:
+                    if i < self.max_attempts - 1:
+                        logger.debug(f"Error querying Redis ({i + 1}/{self.max_attempts}): {ex}")
+                        from time import sleep
+                        sleep(1)
+                        continue
+
+                    else:
+                        from traceback import format_exc
+                        logger.error(f"Failed to query Redis after {self.max_attempts} attempts: {ex}\n{format_exc()}")
+                        raise
+
+        return wrapper
 
 
 ########################################################################################################################