+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Code cleanup
 - Will only update local cache if the results are not empty
 - Added more logging outputs
+- Improved redis request handling
 
 ## 0.3.5
 - Updated to conform with CloudHarvestCoreTasks 0.7.0